Refactoring to move common code b/w CCP and Gateway clients
