+++ internal/fabric/client/client_ccp.go
@@ -28,8 +28,6 @@ import (
 	"github.com/hyperledger/fabric-sdk-go/pkg/fabsdk"
 	mspImpl "github.com/hyperledger/fabric-sdk-go/pkg/msp"
 	"github.com/hyperledger/firefly-fabconnect/internal/errors"
-	eventsapi "github.com/hyperledger/firefly-fabconnect/internal/events/api"
-	"github.com/hyperledger/firefly-fabconnect/internal/fabric/utils"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -42,15 +40,9 @@ type ccpClientWrapper struct {
 }
 
 type ccpRPCWrapper struct {
-	txTimeout           int
-	configProvider      core.ConfigProvider
-	sdk                 *fabsdk.FabricSDK
-	cryptoSuiteConfig   core.CryptoSuiteConfig
-	userStore           msp.UserStore
-	idClient            IdentityClient
-	ledgerClientWrapper *ledgerClientWrapper
-	eventClientWrapper  *eventClientWrapper
-	channelCreator      channelCreator
+	*commonRPCWrapper
+	cryptoSuiteConfig core.CryptoSuiteConfig
+	userStore         msp.UserStore
 	// one channel client per channel ID, per signer ID
 	channelClients map[string](map[string]*ccpClientWrapper)
 }
@@ -69,16 +61,18 @@ func newRPCClientFromCCP(configProvider core.ConfigProvider, txTimeout int, user
 
 	log.Infof("New gRPC connection established")
 	w := &ccpRPCWrapper{
-		sdk:                 ledgerClientWrapper.sdk,
-		configProvider:      configProvider,
-		cryptoSuiteConfig:   cryptoConfig,
-		userStore:           userStore,
-		idClient:            idClient,
-		ledgerClientWrapper: ledgerClientWrapper,
-		eventClientWrapper:  eventClientWrapper,
-		channelClients:      make(map[string]map[string]*ccpClientWrapper),
-		channelCreator:      createChannelClient,
-		txTimeout:           txTimeout,
+		commonRPCWrapper: &commonRPCWrapper{
+			sdk:                 ledgerClientWrapper.sdk,
+			configProvider:      configProvider,
+			idClient:            idClient,
+			ledgerClientWrapper: ledgerClientWrapper,
+			eventClientWrapper:  eventClientWrapper,
+			channelCreator:      createChannelClient,
+			txTimeout:           txTimeout,
+		},
+		cryptoSuiteConfig: cryptoConfig,
+		userStore:         userStore,
+		channelClients:    make(map[string]map[string]*ccpClientWrapper),
 	}
 	return w, nil
 }
@@ -126,59 +120,6 @@ func (w *ccpRPCWrapper) Query(channelId, signer, chaincodeName, method string, a
 
 	log.Tracef("RPC [%s:%s:%s] <-- %+v", channelId, chaincodeName, method, result)
 	return result.Payload, nil
-}
-
-func (w *ccpRPCWrapper) QueryTransaction(channelId, signer, txId string) (map[string]interface{}, error) {
-	log.Tracef("RPC [%s] --> QueryTransaction %s", channelId, txId)
-
-	result, err := w.ledgerClientWrapper.queryTransaction(channelId, signer, txId)
-	if err != nil {
-		log.Errorf("Failed to query transaction on channel %s. %s", channelId, err)
-		return nil, err
-	}
-
-	log.Tracef("RPC [%s] <-- %+v", channelId, result)
-	return result, nil
-}
-
-func (w *ccpRPCWrapper) QueryChainInfo(channelId, signer string) (*fab.BlockchainInfoResponse, error) {
-	log.Tracef("RPC [%s] --> ChainInfo", channelId)
-
-	result, err := w.ledgerClientWrapper.queryChainInfo(channelId, signer)
-	if err != nil {
-		log.Errorf("Failed to query chain info on channel %s. %s", channelId, err)
-		return nil, err
-	}
-
-	log.Tracef("RPC [%s] <-- %+v", channelId, result)
-	return result, nil
-}
-
-func (w *ccpRPCWrapper) QueryBlock(channelId string, blockNumber uint64, signer string) (*utils.RawBlock, *utils.Block, error) {
-	log.Tracef("RPC [%s] --> QueryBlock %v", channelId, blockNumber)
-
-	rawblock, block, err := w.ledgerClientWrapper.queryBlock(channelId, blockNumber, signer)
-	if err != nil {
-		log.Errorf("Failed to query block %v on channel %s. %s", blockNumber, channelId, err)
-		return nil, nil, err
-	}
-
-	log.Tracef("RPC [%s] <-- success", channelId)
-	return rawblock, block, nil
-}
-
-// The returned registration must be closed when done
-func (w *ccpRPCWrapper) SubscribeEvent(subInfo *eventsapi.SubscriptionInfo, since uint64) (*RegistrationWrapper, <-chan *fab.BlockEvent, <-chan *fab.CCEvent, error) {
-	reg, blockEventCh, ccEventCh, err := w.eventClientWrapper.subscribeEvent(subInfo, since)
-	if err != nil {
-		log.Errorf("Failed to subscribe to event [%s:%s:%s]. %s", subInfo.Stream, subInfo.ChannelId, subInfo.Filter.ChaincodeId, err)
-		return nil, nil, nil, err
-	}
-	return reg, blockEventCh, ccEventCh, nil
-}
-
-func (w *ccpRPCWrapper) Unregister(regWrapper *RegistrationWrapper) {
-	regWrapper.eventClient.Unregister(regWrapper.registration)
 }
 
 func (w *ccpRPCWrapper) getChannelClient(channelId string, signer string) (*ccpClientWrapper, error) {
@@ -247,23 +188,4 @@ func (w *ccpRPCWrapper) sendTransaction(channelId, signer, chaincodeName, method
 		return nil, nil, nil, err
 	}
 	return client.signer, result.Payload, &txStatus, nil
-}
-
-func convert(args []string) [][]byte {
-	result := [][]byte{}
-	for _, v := range args {
-		result = append(result, []byte(v))
-	}
-	return result
-}
-
-func newReceipt(responsePayload []byte, status *fab.TxStatusEvent, signerID *msp.IdentityIdentifier) *TxReceipt {
-	return &TxReceipt{
-		SignerMSP:     signerID.MSPID,
-		Signer:        signerID.ID,
-		TransactionID: status.TxID,
-		Status:        status.TxValidationCode,
-		BlockNumber:   status.BlockNumber,
-		SourcePeer:    status.SourceURL,
-	}
 }
