+++ internal/fabric/client/client_common.go
@@ -6,8 +6,24 @@ import (
 	"github.com/hyperledger/fabric-sdk-go/pkg/client/channel"
 	"github.com/hyperledger/fabric-sdk-go/pkg/common/providers/context"
 	"github.com/hyperledger/fabric-sdk-go/pkg/common/providers/core"
+	"github.com/hyperledger/fabric-sdk-go/pkg/common/providers/fab"
+	"github.com/hyperledger/fabric-sdk-go/pkg/common/providers/msp"
+	"github.com/hyperledger/fabric-sdk-go/pkg/fabsdk"
 	"github.com/hyperledger/firefly-fabconnect/internal/errors"
+	eventsapi "github.com/hyperledger/firefly-fabconnect/internal/events/api"
+	"github.com/hyperledger/firefly-fabconnect/internal/fabric/utils"
+	log "github.com/sirupsen/logrus"
 )
+
+type commonRPCWrapper struct {
+	txTimeout           int
+	configProvider      core.ConfigProvider
+	sdk                 *fabsdk.FabricSDK
+	idClient            IdentityClient
+	ledgerClientWrapper *ledgerClientWrapper
+	eventClientWrapper  *eventClientWrapper
+	channelCreator      channelCreator
+}
 
 func getOrgFromConfig(config core.ConfigProvider) (string, error) {
 	configBackend, err := config()
@@ -50,4 +66,76 @@ type channelCreator func(context.ChannelProvider) (*channel.Client, error)
 
 func createChannelClient(channelProvider context.ChannelProvider) (*channel.Client, error) {
 	return channel.New(channelProvider)
+}
+
+func newReceipt(responsePayload []byte, status *fab.TxStatusEvent, signerID *msp.IdentityIdentifier) *TxReceipt {
+	return &TxReceipt{
+		SignerMSP:     signerID.MSPID,
+		Signer:        signerID.ID,
+		TransactionID: status.TxID,
+		Status:        status.TxValidationCode,
+		BlockNumber:   status.BlockNumber,
+		SourcePeer:    status.SourceURL,
+	}
+}
+
+func convert(args []string) [][]byte {
+	result := [][]byte{}
+	for _, v := range args {
+		result = append(result, []byte(v))
+	}
+	return result
+}
+
+func (w *commonRPCWrapper) QueryChainInfo(channelId, signer string) (*fab.BlockchainInfoResponse, error) {
+	log.Tracef("RPC [%s] --> ChainInfo", channelId)
+
+	result, err := w.ledgerClientWrapper.queryChainInfo(channelId, signer)
+	if err != nil {
+		log.Errorf("Failed to query chain info on channel %s. %s", channelId, err)
+		return nil, err
+	}
+
+	log.Tracef("RPC [%s] <-- %+v", channelId, result)
+	return result, nil
+}
+
+func (w *commonRPCWrapper) QueryBlock(channelId string, blockNumber uint64, signer string) (*utils.RawBlock, *utils.Block, error) {
+	log.Tracef("RPC [%s] --> QueryBlock %v", channelId, blockNumber)
+
+	rawblock, block, err := w.ledgerClientWrapper.queryBlock(channelId, blockNumber, signer)
+	if err != nil {
+		log.Errorf("Failed to query block %v on channel %s. %s", blockNumber, channelId, err)
+		return nil, nil, err
+	}
+
+	log.Tracef("RPC [%s] <-- success", channelId)
+	return rawblock, block, nil
+}
+
+func (w *commonRPCWrapper) QueryTransaction(channelId, signer, txId string) (map[string]interface{}, error) {
+	log.Tracef("RPC [%s] --> QueryTransaction %s", channelId, txId)
+
+	result, err := w.ledgerClientWrapper.queryTransaction(channelId, signer, txId)
+	if err != nil {
+		log.Errorf("Failed to query transaction on channel %s. %s", channelId, err)
+		return nil, err
+	}
+
+	log.Tracef("RPC [%s] <-- %+v", channelId, result)
+	return result, nil
+}
+
+// The returned registration must be closed when done
+func (w *commonRPCWrapper) SubscribeEvent(subInfo *eventsapi.SubscriptionInfo, since uint64) (*RegistrationWrapper, <-chan *fab.BlockEvent, <-chan *fab.CCEvent, error) {
+	reg, blockEventCh, ccEventCh, err := w.eventClientWrapper.subscribeEvent(subInfo, since)
+	if err != nil {
+		log.Errorf("Failed to subscribe to event [%s:%s:%s]. %s", subInfo.Stream, subInfo.ChannelId, subInfo.Filter.ChaincodeId, err)
+		return nil, nil, nil, err
+	}
+	return reg, blockEventCh, ccEventCh, nil
+}
+
+func (w *commonRPCWrapper) Unregister(regWrapper *RegistrationWrapper) {
+	regWrapper.eventClient.Unregister(regWrapper.registration)
 }
