+++ internal/fabric/client/client_gateway_clientside.go
@@ -27,8 +27,6 @@ import (
 	"github.com/hyperledger/fabric-sdk-go/pkg/fabsdk"
 	"github.com/hyperledger/fabric-sdk-go/pkg/gateway"
 	"github.com/hyperledger/firefly-fabconnect/internal/errors"
-	eventsapi "github.com/hyperledger/firefly-fabconnect/internal/events/api"
-	"github.com/hyperledger/firefly-fabconnect/internal/fabric/utils"
 	log "github.com/sirupsen/logrus"
 )
 
@@ -37,14 +35,9 @@ type gatewayCreator func(core.ConfigProvider, string, int) (*gateway.Gateway, er
 type networkCreator func(*gateway.Gateway, string) (*gateway.Network, error)
 
 type gwRPCWrapper struct {
-	txTimeout           int
-	configProvider      core.ConfigProvider
-	idClient            IdentityClient
-	ledgerClientWrapper *ledgerClientWrapper
-	eventClientWrapper  *eventClientWrapper
-	gatewayCreator      gatewayCreator
-	networkCreator      networkCreator
-	channelCreator      channelCreator
+	*commonRPCWrapper
+	gatewayCreator gatewayCreator
+	networkCreator networkCreator
 	// networkCreator networkC
 	// one gateway client per signer
 	gwClients map[string]*gateway.Gateway
@@ -56,17 +49,19 @@ type gwRPCWrapper struct {
 
 func newRPCClientWithClientSideGateway(configProvider core.ConfigProvider, txTimeout int, idClient IdentityClient, ledgerClientWrapper *ledgerClientWrapper, eventClientWrapper *eventClientWrapper) (RPCClient, error) {
 	return &gwRPCWrapper{
-		txTimeout:           txTimeout,
-		configProvider:      configProvider,
-		idClient:            idClient,
-		ledgerClientWrapper: ledgerClientWrapper,
-		eventClientWrapper:  eventClientWrapper,
-		gatewayCreator:      createGateway,
-		networkCreator:      getNetwork,
-		channelCreator:      createChannelClient,
-		gwClients:           make(map[string]*gateway.Gateway),
-		gwGatewayClients:    make(map[string]map[string]*gateway.Network),
-		gwChannelClients:    make(map[string]map[string]*channel.Client),
+		commonRPCWrapper: &commonRPCWrapper{
+			txTimeout:           txTimeout,
+			configProvider:      configProvider,
+			idClient:            idClient,
+			ledgerClientWrapper: ledgerClientWrapper,
+			eventClientWrapper:  eventClientWrapper,
+			channelCreator:      createChannelClient,
+		},
+		gatewayCreator:   createGateway,
+		networkCreator:   getNetwork,
+		gwClients:        make(map[string]*gateway.Gateway),
+		gwGatewayClients: make(map[string]map[string]*gateway.Network),
+		gwChannelClients: make(map[string]map[string]*channel.Client),
 	}, nil
 }
 
@@ -100,10 +95,7 @@ func (w *gwRPCWrapper) Query(channelId, signer, chaincodeName, method string, ar
 		return nil, err
 	}
 
-	bytes := make([][]byte, len(args))
-	for i, v := range args {
-		bytes[i] = []byte(v)
-	}
+	bytes := convert(args)
 	req := channel.Request{
 		ChaincodeID: chaincodeName,
 		Fcn:         method,
@@ -117,59 +109,6 @@ func (w *gwRPCWrapper) Query(channelId, signer, chaincodeName, method string, ar
 
 	log.Tracef("RPC [%s:%s:%s] <-- %+v", channelId, chaincodeName, method, result)
 	return result.Payload, nil
-}
-
-func (w *gwRPCWrapper) QueryChainInfo(channelId, signer string) (*fab.BlockchainInfoResponse, error) {
-	log.Tracef("RPC [%s] --> ChainInfo", channelId)
-
-	result, err := w.ledgerClientWrapper.queryChainInfo(channelId, signer)
-	if err != nil {
-		log.Errorf("Failed to query chain info on channel %s. %s", channelId, err)
-		return nil, err
-	}
-
-	log.Tracef("RPC [%s] <-- %+v", channelId, result)
-	return result, nil
-}
-
-func (w *gwRPCWrapper) QueryBlock(channelId string, blockNumber uint64, signer string) (*utils.RawBlock, *utils.Block, error) {
-	log.Tracef("RPC [%s] --> QueryBlock %v", channelId, blockNumber)
-
-	rawblock, block, err := w.ledgerClientWrapper.queryBlock(channelId, blockNumber, signer)
-	if err != nil {
-		log.Errorf("Failed to query block %v on channel %s. %s", blockNumber, channelId, err)
-		return nil, nil, err
-	}
-
-	log.Tracef("RPC [%s] <-- success", channelId)
-	return rawblock, block, nil
-}
-
-func (w *gwRPCWrapper) QueryTransaction(channelId, signer, txId string) (map[string]interface{}, error) {
-	log.Tracef("RPC [%s] --> QueryTransaction %s", channelId, txId)
-
-	result, err := w.ledgerClientWrapper.queryTransaction(channelId, signer, txId)
-	if err != nil {
-		log.Errorf("Failed to query transaction on channel %s. %s", channelId, err)
-		return nil, err
-	}
-
-	log.Tracef("RPC [%s] <-- %+v", channelId, result)
-	return result, nil
-}
-
-// The returned registration must be closed when done
-func (w *gwRPCWrapper) SubscribeEvent(subInfo *eventsapi.SubscriptionInfo, since uint64) (*RegistrationWrapper, <-chan *fab.BlockEvent, <-chan *fab.CCEvent, error) {
-	reg, blockEventCh, ccEventCh, err := w.eventClientWrapper.subscribeEvent(subInfo, since)
-	if err != nil {
-		log.Errorf("Failed to subscribe to event [%s:%s:%s]. %s", subInfo.Stream, subInfo.ChannelId, subInfo.Filter.ChaincodeId, err)
-		return nil, nil, nil, err
-	}
-	return reg, blockEventCh, ccEventCh, nil
-}
-
-func (w *gwRPCWrapper) Unregister(regWrapper *RegistrationWrapper) {
-	regWrapper.eventClient.Unregister(regWrapper.registration)
 }
 
 func (w *gwRPCWrapper) Close() error {
